refactor: consolidate DynamoDB client dependencies
Remove duplicate get_dynamodb_client implementations across routers
and centralize them in api/dependencies.py for DRY principle.

Changes:
- Added get_dynamodb_client() to dependencies.py (required context)
- Added get_dynamodb_client_optional() to dependencies.py (optional context)
- Updated all routers to import from dependencies instead of duplicating
- Removed 4 duplicate implementations across routers

Benefits:
- Single source of truth for dependency injection
- Easier to maintain and update AWS context validation logic
- Consistent error handling across all endpoints

Related to #451
