+++ api/routers/tradingview.py
@@ -1,10 +1,11 @@
 from fastapi import APIRouter, Depends, HTTPException
 
+from api.dependencies import get_dynamodb_client
 from api.models.tradingview import (
     SetTradingViewLinkRequest,
     SetTradingViewLinkResponse,
 )
-from client.aws_client import AwsClient, DynamoDBClient
+from client.aws_client import DynamoDBClient
 from utils.logger import Logger
 
 router = APIRouter(
@@ -12,20 +13,6 @@ router = APIRouter(
     tags=["tradingview"],
 )
 logger = Logger.get_logger("tradingview_router")
-
-
-def get_dynamodb_client() -> DynamoDBClient:
-    """
-    Create DynamoDBClient instance.
-    Validates AWS context before allowing access.
-    """
-    if not AwsClient.is_aws_context():
-        raise HTTPException(
-            status_code=403,
-            detail="AWS context not available. "
-            "Set AWS_PROFILE environment variable.",
-        )
-    return DynamoDBClient()
 
 
 @router.put("", response_model=SetTradingViewLinkResponse)
